+++ lib/templates/commands/cloneTemplate.ts
@@ -1,10 +1,10 @@
 import { prisma, prismaErrors, Prisma } from "@gcforms/database";
 import { FormRecord } from "@lib/types";
 
-import { authorization } from "../privileges";
-import { AuditLogAccessDeniedDetails, AuditLogDetails, logEvent } from "../auditLogs";
+import { authorization } from "../../privileges";
+import { AuditLogAccessDeniedDetails, AuditLogDetails, logEvent } from "../../auditLogs";
 import { logMessage } from "@lib/logger";
-import { parseTemplate } from "./internal";
+import { parseTemplate } from "../internal";
 
 /**
  * Clone a template including associated users and delivery option
