+++ lib/templates/commands/removeDeliveryOption.ts
@@ -1,8 +1,8 @@
 import { prisma } from "@gcforms/database";
 
-import { authorization } from "../privileges";
-import { AuditLogAccessDeniedDetails, AuditLogDetails, logEvent } from "../auditLogs";
-import { TemplateAlreadyPublishedError, TemplateNotFoundError } from "./internal/errors";
+import { authorization } from "../../privileges";
+import { AuditLogAccessDeniedDetails, AuditLogDetails, logEvent } from "../../auditLogs";
+import { TemplateAlreadyPublishedError, TemplateNotFoundError } from "../internal/errors";
 
 /**
  * Remove DeliveryOption from template. Form responses will be sent to the Vault.
