+++ lib/templates/commands/restoreTemplate.ts
@@ -1,11 +1,11 @@
-import { formCache } from "../cache/formCache";
+import { formCache } from "../../cache/formCache";
 import { prisma, prismaErrors } from "@gcforms/database";
 import { FormRecord } from "@lib/types";
 
-import { authorization } from "../privileges";
-import { AuditLogAccessDeniedDetails, AuditLogEvent, logEvent } from "../auditLogs";
-import { TemplateNotFoundError } from "./internal/errors";
-import { parseTemplate } from "./internal";
+import { authorization } from "../../privileges";
+import { AuditLogAccessDeniedDetails, AuditLogEvent, logEvent } from "../../auditLogs";
+import { TemplateNotFoundError } from "../internal/errors";
+import { parseTemplate } from "../internal";
 
 /**
  * Restores a form template from the archived state.
