+++ lib/templates/commands/syncAssignedUsersForTemplate.ts
@@ -1,12 +1,12 @@
 import { prisma, prismaErrors } from "@gcforms/database";
 import { FormRecord, FormProperties } from "@lib/types";
 
-import { authorization } from "../privileges";
-import { AuditLogAccessDeniedDetails, AuditLogDetails, logEvent } from "../auditLogs";
+import { authorization } from "../../privileges";
+import { AuditLogAccessDeniedDetails, AuditLogDetails, logEvent } from "../../auditLogs";
 import { logMessage } from "@lib/logger";
-import { invalidateTemplateEditLockUserCountCache } from "../editLocks";
-import { notifyOwnersOwnerAdded, notifyOwnersOwnerRemoved } from "./internal/notifications";
-import { parseTemplate } from "./internal";
+import { invalidateTemplateEditLockUserCountCache } from "../../editLocks";
+import { notifyOwnersOwnerAdded, notifyOwnersOwnerRemoved } from "../internal/notifications";
+import { parseTemplate } from "../internal";
 
 /**
  * Add/remove (sync) users to a form
