+++ lib/templates/commands/updateFormPurpose.ts
@@ -1,10 +1,10 @@
 import { prisma, prismaErrors, Prisma } from "@gcforms/database";
 import { FormRecord } from "@lib/types";
 
-import { authorization } from "../privileges";
-import { AuditLogAccessDeniedDetails, AuditLogDetails, logEvent } from "../auditLogs";
-import { TemplateAlreadyPublishedError } from "./internal/errors";
-import { parseTemplate } from "./internal";
+import { authorization } from "../../privileges";
+import { AuditLogAccessDeniedDetails, AuditLogDetails, logEvent } from "../../auditLogs";
+import { TemplateAlreadyPublishedError } from "../internal/errors";
+import { parseTemplate } from "../internal";
 
 export async function updateFormPurpose(
   formID: string,
