+++ lib/templates/guards/assertCanEditTemplate.ts
