+++ lib/templates/mappers/mapTemplateToPublicFormRecord.ts
