+++ lib/templates/queries/getTemplatePublishedStatus.ts
